+++ packages/wiz-ui-react/src/components/base/inputs/text/components/index.ts
@@ -1,1 +1,1 @@
-export { WizTextInput } from "./text";
+export { WizTextInput } from "./text-input";
